src/third_party/blink/web_tests/http/tests/devtools/elements/styles-1/css-live-edit-expected.txt

Tests that styles are updated when live-editing css resource.


Running: testLiveEdit
color: rgb(0, 0, 0);
    -internal-root-color - html user agent stylesheet
display: block;
    block - div user agent stylesheet
font-size: 20px;
    20px - body css-live-edit.css:1 -> css-live-edit.css:1:7
[expanded] 
element.style { ()
